the last test was passed

+++ include/lmodules.h
@@ -5,7 +5,7 @@ using std::vector;
 #ifndef LMODULES_H
 #define LMODULES_H
 
-vector<double> get_gradation(int, int, vector<vector<double>>, vector<int>);
+vector<double> get_gradation(int, int, vector<vector<double>>, vector<double>);
 vector<int> get_free_vars();
 
 #endif
